requirements.txt
# Trading Bot - Production Ready (Option A Architecture)
sqlalchemy>=2.0.0
google-generativeai>=0.3.0
python-telegram-bot>=20.0
tweepy>=4.14.0
aiohttp>=3.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
